Slight refactorization
