+++ DMTools.Blazor/Components/Pages/StatBlock/Languages.razor
@@ -5,5 +5,5 @@
 
 @code {
     [Parameter]
-    public DMToolsLibrary.StatBlocks.StatBlockHelpers.Languages languages { get; set; }
+    public DMTools.StatBlocks.StatBlockHelpers.Languages languages { get; set; }
 }
