+++ DMTools.Blazor/Components/Pages/StatBlock/Senses.razor
@@ -5,5 +5,5 @@
 
 @code {
 	[Parameter]
-	public DMToolsLibrary.StatBlocks.StatBlockHelpers.Senses senses { get; set; }
+	public DMTools.StatBlocks.StatBlockHelpers.Senses senses { get; set; }
 }
