+++ DMTools.Blazor/Components/_Imports.razor
@@ -8,10 +8,11 @@
 @using Microsoft.JSInterop
 @using DMTools.Blazor
 @using DMTools.Blazor.Components
-@using DMToolsLibrary;
-@using DMToolsLibrary.StatBlocks.Types;
-@using DMToolsLibrary.StatBlocks.StatBlockHelpers
-@using static DMToolsLibrary.Enums.LibraryEnums;
+@using DMTools;
+@using DMTools.StatBlocks.Types;
+@using DMTools.StatBlocks.StatBlockHelpers
+@using static DMTools.Shared.Enums.LibraryEnums;
+@using DMTools.Shared.Enums;
 @using System.Text.RegularExpressions;
 @using DMTools.Database.Entities;
-@using DMToolsLibrary.Security;
+@using DMTools.Security;
