+++ DMTools.NameGenerator/DMTools.NameGenerator.csproj
@@ -1,18 +1,14 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <OutputType>Exe</OutputType>
     <TargetFramework>net8.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\DMToolsLibrary\DMTools.StatBlock.csproj" />
+    <ProjectReference Include="..\DMTools.Database\DMTools.Database.csproj" />
+    <ProjectReference Include="..\DMTools.Shared\DMTools.Shared.csproj" />
   </ItemGroup>
 
 </Project>
