+++ DMTools.ScriptTracker/Class1.cs
@@ -1,0 +1,7 @@
+﻿namespace DMTools.ScriptTracker
+{
+    public class Class1
+    {
+
+    }
+}
