+++ DMTools.sln
@@ -3,13 +3,23 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.8.34330.188
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DMToolsLibrary", "DMToolsLibrary\DMToolsLibrary.csproj", "{6F115A74-81FA-41D7-ABD1-F3A2CFA60387}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DMTools.StatBlock", "DMToolsLibrary\DMTools.StatBlock.csproj", "{6F115A74-81FA-41D7-ABD1-F3A2CFA60387}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DMToolsTest", "DMToolsTest\DMToolsTest.csproj", "{10BFEB88-0DC5-40DB-BDA4-40DC3DA435E0}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DMTools.Blazor", "DMTools.Blazor\DMTools.Blazor.csproj", "{AD684E2E-197A-447C-A4A1-5FD72D285C58}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DMTools.Database", "DMTools.Database\DMTools.Database.csproj", "{4CBA827B-DA69-4809-922D-7CB1E6464297}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DMTools.Security", "DMTools.Security\DMTools.Security.csproj", "{A051FF8E-2300-4FA0-9A2C-460300D7A911}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DMTools.ScriptTracker", "DMTools.ScriptTracker\DMTools.ScriptTracker.csproj", "{CA37E87B-DD70-4698-85D6-DFB348E48C4F}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DMTools.EncounterGenerator", "DMTools.EncounterGenerator\DMTools.EncounterGenerator.csproj", "{2154708C-AD25-48BD-BD15-C458F10148C9}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DMTools.Shared", "DMTools.Shared\DMTools.Shared.csproj", "{4F02AB3F-541D-4E93-9481-DE98D0150720}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DMTools.NameGenerator", "DMTools.NameGenerator\DMTools.NameGenerator.csproj", "{973ABD7B-01A2-4F6B-ADFC-BB5D86CCBCCB}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -33,6 +43,26 @@ Global
 		{4CBA827B-DA69-4809-922D-7CB1E6464297}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{4CBA827B-DA69-4809-922D-7CB1E6464297}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{4CBA827B-DA69-4809-922D-7CB1E6464297}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A051FF8E-2300-4FA0-9A2C-460300D7A911}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A051FF8E-2300-4FA0-9A2C-460300D7A911}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A051FF8E-2300-4FA0-9A2C-460300D7A911}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A051FF8E-2300-4FA0-9A2C-460300D7A911}.Release|Any CPU.Build.0 = Release|Any CPU
+		{CA37E87B-DD70-4698-85D6-DFB348E48C4F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{CA37E87B-DD70-4698-85D6-DFB348E48C4F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{CA37E87B-DD70-4698-85D6-DFB348E48C4F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{CA37E87B-DD70-4698-85D6-DFB348E48C4F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{2154708C-AD25-48BD-BD15-C458F10148C9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{2154708C-AD25-48BD-BD15-C458F10148C9}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{2154708C-AD25-48BD-BD15-C458F10148C9}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{2154708C-AD25-48BD-BD15-C458F10148C9}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4F02AB3F-541D-4E93-9481-DE98D0150720}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4F02AB3F-541D-4E93-9481-DE98D0150720}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4F02AB3F-541D-4E93-9481-DE98D0150720}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4F02AB3F-541D-4E93-9481-DE98D0150720}.Release|Any CPU.Build.0 = Release|Any CPU
+		{973ABD7B-01A2-4F6B-ADFC-BB5D86CCBCCB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{973ABD7B-01A2-4F6B-ADFC-BB5D86CCBCCB}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{973ABD7B-01A2-4F6B-ADFC-BB5D86CCBCCB}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{973ABD7B-01A2-4F6B-ADFC-BB5D86CCBCCB}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
