+++ DMToolsLibrary/DiceCalculator/DiceFunctions.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace DMToolsLibrary.DiceCalculator;
+namespace DMTools.DiceCalculator;
 
 public static class DiceFunctions
 {
