+++ DMToolsLibrary/DiceCalculator/HitDie.cs
@@ -4,9 +4,9 @@ using System.Drawing;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using static DMToolsLibrary.Enums.LibraryEnums;
+using static DMTools.Shared.Enums.LibraryEnums;
 
-namespace DMToolsLibrary.DiceCalculator
+namespace DMTools.DiceCalculator
 {
     public class HitDie
     {
