+++ DMToolsLibrary/StatBlocks/StatBlockHelpers/Speeds.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace DMToolsLibrary.StatBlocks.StatBlockHelpers
+namespace DMTools.StatBlocks.StatBlockHelpers
 {
     public class Speeds
     {
